# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: |
    Omninity Operator HTTP API.

    Every 2xx response conforms to the canonical envelope shape defined in
    `Standard 1` of the bug-prevention standards: `{ success, data, error }`.
    Collection endpoints further wrap their payload in
    `{ items, nextCursor }` per `Standard 13`.

    Tenant context is supplied via the `X-Tenant-ID` header until session
    cookies replace it on `/api/auth/login`.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /models:
    get:
      operationId: listModels
      tags: [models]
      summary: List Ollama models known to this installation
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Models page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelListResponse"
  /models/pull:
    post:
      operationId: pullModel
      tags: [models]
      summary: Schedule download of a named model
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelPullRequest"
      responses:
        "200":
          description: Pull scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPullResponse"
  /models/{name}:
    get:
      operationId: getModel
      tags: [models]
      summary: Fetch one model's status (singleton)
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Model details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelGetResponse"
        "404":
          description: Unknown model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
  /tools:
    get:
      operationId: listTools
      tags: [tools]
      summary: List installed tools
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Tools page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolListResponse"
  /tools/{name}/invoke:
    post:
      operationId: invokeTool
      tags: [tools]
      summary: Directly invoke a tool (outside an agent run)
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolInvokeRequest"
      responses:
        "200":
          description: Tool output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolInvokeResponse"
  /privacy/events:
    get:
      operationId: listPrivacyEvents
      tags: [privacy]
      summary: Browse the privacy-event audit log
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Privacy events page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivacyEventListResponse"
    post:
      operationId: createPrivacyEvent
      tags: [privacy]
      summary: Manually log a privacy event
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrivacyEventRequest"
      responses:
        "200":
          description: Event recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivacyEventResponse"
  /memory:
    get:
      operationId: listMemories
      tags: [memory]
      summary: List long-lived user memories
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Memories page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryListResponse"
    post:
      operationId: createMemory
      tags: [memory]
      summary: Create a memory entry
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemoryRequest"
      responses:
        "200":
          description: Memory created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryResponse"
  /memory/{id}:
    get:
      operationId: getMemory
      tags: [memory]
      summary: Fetch one memory (singleton)
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Memory entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryResponse"
        "404":
          description: Unknown memory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    delete:
      operationId: deleteMemory
      tags: [memory]
      summary: Delete a memory entry
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deletion receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryDeleteResponse"
  /files/list:
    get:
      operationId: listFiles
      tags: [files]
      summary: List entries inside the workspace sandbox
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: path
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Files page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileListResponse"
  /files/read:
    post:
      operationId: readFile
      tags: [files]
      summary: Read a sandboxed file
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileReadRequest"
      responses:
        "200":
          description: File contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileReadResponse"
  /files/write:
    post:
      operationId: writeFile
      tags: [files]
      summary: Write a sandboxed file
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileWriteRequest"
      responses:
        "200":
          description: Write receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileWriteResponse"
  /files/delete:
    post:
      operationId: deleteFile
      tags: [files]
      summary: Delete a sandboxed file
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileDeleteRequest"
      responses:
        "200":
          description: Delete receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileDeleteResponse"
...
components:
  parameters:
    TenantHeader:
      name: X-Tenant-ID
      in: header
      required: true
      description: |
        Tenant identifier. Replaced by JWT-derived context once full SSO
        ships — until then this header is the request's tenant context.
      schema:
        type: string
    CursorParam:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned by the previous page.
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      required: false
      description: Page size, default 20, max 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ApiErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          $ref: "#/components/schemas/ApiError"
...
    ModelEntry:
      type: object
      required: [name, status]
      properties:
        name:
          type: string
        status:
          type: string
        sizeBytes:
          type: integer
          nullable: true
        family:
          type: string
          nullable: true
        modifiedAt:
          type: string
          format: date-time
          nullable: true
    ModelListPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ModelEntry"
        nextCursor:
          type: string
          nullable: true
    ModelListResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/ModelListPage"
    ModelGetResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/ModelEntry"
    ModelPullRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    ModelPullReceipt:
      type: object
      required: [name, status, scheduledAt]
      properties:
        name:
          type: string
        status:
          type: string
        scheduledAt:
          type: string
          format: date-time
    ModelPullResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/ModelPullReceipt"
...
    ToolEntry:
      type: object
      required: [name, description, riskLevel]
      properties:
        name:
          type: string
        description:
          type: string
        riskLevel:
          type: string
    ToolListPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ToolEntry"
        nextCursor:
          type: string
          nullable: true
    ToolListResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/ToolListPage"
    ToolInvokeRequest:
      type: object
      required: [input]
      properties:
        input:
          type: object
          additionalProperties: true
    ToolInvokeResult:
      type: object
      required: [toolName, output, durationMs]
      properties:
        toolName:
          type: string
        output:
          type: object
          additionalProperties: true
        durationMs:
          type: integer
    ToolInvokeResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/ToolInvokeResult"
    PrivacyEvent:
      type: object
      required: [id, eventType, actor, target, severity, createdAt]
      properties:
        id:
          type: string
        eventType:
          type: string
        actor:
          type: string
        target:
          type: string
        severity:
          type: string
        detail:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    PrivacyEventListPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PrivacyEvent"
        nextCursor:
          type: string
          nullable: true
    PrivacyEventListResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/PrivacyEventListPage"
    PrivacyEventResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/PrivacyEvent"
    CreatePrivacyEventRequest:
      type: object
      required: [eventType, actor, target]
      properties:
        eventType:
          type: string
        actor:
          type: string
        target:
          type: string
        severity:
          type: string
        detail:
          type: string
    Memory:
      type: object
      required: [id, kind, title, content, importance, createdAt, updatedAt]
      properties:
        id:
          type: string
        kind:
          type: string
        title:
          type: string
        content:
          type: string
        importance:
          type: integer
        source:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MemoryListPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Memory"
        nextCursor:
          type: string
          nullable: true
    MemoryListResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/MemoryListPage"
    MemoryResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/Memory"
    CreateMemoryRequest:
      type: object
      required: [title, content]
      properties:
        kind:
          type: string
        title:
          type: string
        content:
          type: string
        importance:
          type: integer
          minimum: 0
          maximum: 100
        source:
          type: string
    MemoryDeleteReceipt:
      type: object
      required: [id, deleted]
      properties:
        id:
          type: string
        deleted:
          type: boolean
    MemoryDeleteResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/MemoryDeleteReceipt"
    FileEntry:
      type: object
      required: [path, kind, size]
      properties:
        path:
          type: string
        kind:
          type: string
        size:
          type: integer
        modifiedAt:
          type: string
          format: date-time
          nullable: true
    FileListPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FileEntry"
        nextCursor:
          type: string
          nullable: true
    FileListResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/FileListPage"
    FileReadRequest:
      type: object
      required: [path]
      properties:
        path:
          type: string
    FileReadResult:
      type: object
      required: [path, content, size]
      properties:
        path:
          type: string
        content:
          type: string
        size:
          type: integer
    FileReadResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/FileReadResult"
    FileWriteRequest:
      type: object
      required: [path, content]
      properties:
        path:
          type: string
        content:
          type: string
    FileWriteResult:
      type: object
      required: [path, size]
      properties:
        path:
          type: string
        size:
          type: integer
    FileWriteResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/FileWriteResult"
    FileDeleteRequest:
      type: object
      required: [path]
      properties:
        path:
          type: string
    FileDeleteResult:
      type: object
      required: [path, deleted]
      properties:
        path:
          type: string
        deleted:
          type: boolean
    FileDeleteResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/FileDeleteResult"
...